requirements.txt
fastapi
uvicorn
torch
transformers
faiss-cpu
python-dotenv
jinja2
